[DEV] Feat/seo (#32)
* chore: Update npm

* feat(seo): Add seo in layout root

* refactor(auth): Add seo and make auth page static

* refactor(auth): Separate client component and create page's component in order to SEO work as well

* refactor(auth): Remove use client in order to fill metadata

* feat(seo): Add seo in root dashboard layout

* feat(seo): Integrate seo in landing page

* chore(auth): Modify wrong import useRouter in loginform page's component

* refactor(chatbot): Add seo to chatbot and make the page static

* chore: Remove duplicated line

* feat: Add favicon

* feat: Add og image

* feat(seo): Add robots file

* feat(seo): Add sitemap for seo optimation

* feat(seo): Add keyword in root layout

* fix(card): Fix dashboard card show 4 item on resolution 1300px

* refactor(seo): Update all the pages into static pages by separating it to pages component
