+++ app/layout.tsx
@@ -14,9 +14,33 @@ const poppins = Poppins({
 })
 
 export const metadata: Metadata = {
-  title: 'Melancong',
+  title: {
+    default: 'Melancong - Tourism Recommendation in Bali',
+    template: '%s | Melancong - Tourism Recommendation in Bali'
+  },
+  keywords: 'melancong, travel, recommendation, bali',
   description:
-    'Melancong adalah suatu website yang dimana menampilkan destinasi wisata yang berada di bali, memberikan rekomendasi melancong sekitaran bali dengan berbagai fitur dan memberikan rekomendasi terbaik untuk menentukan decision making'
+    'Melancong is a website that showcases tourist destinations in Bali, provides travel recommendations around Bali with various features, and offers the best recommendations to aid in decision making',
+  openGraph: {
+    title: 'Melancong - Tourism Recommendation in Bali',
+    description:
+      'Melancong is a website that showcases tourist destinations in Bali, provides travel recommendations around Bali with various features, and offers the best recommendations to aid in decision making',
+    type: 'website',
+    locale: 'en_US',
+    url: 'https://melancong-fe.vercel.app/',
+    siteName: 'Melancong'
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true
+    }
+  },
+  icons: {
+    icon: '/favicon.ico'
+  }
 }
 
 export default function RootLayout({
@@ -29,9 +53,7 @@ export default function RootLayout({
       <body className={cn('min-h-dvh font-sans antialiased', poppins.variable)}>
         <NextUIProvider>
           <ThemeProvider attribute='class' defaultTheme='light'>
-            {/* <header></header> */}
             <main>{children}</main>
-            {/* <footer></footer> */}
           </ThemeProvider>
         </NextUIProvider>
       </body>
