+++ app/robots.ts
@@ -1,0 +1,12 @@
+import { MetadataRoute } from 'next'
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: '*',
+      allow: '/',
+      disallow: '/private/'
+    },
+    sitemap: 'https://melancong-fe.vercel.app/sitemap.xml'
+  }
+}
